+++ onyx-database/src/main/kotlin/com/onyx/diskmap/store/impl/EncryptedMemoryMappedStore.kt
@@ -19,7 +19,7 @@ class EncryptedMemoryMappedStore (filePath: String, context: SchemaContext, dele
      * @since 2.2.0
      */
     @Suppress("UNCHECKED_CAST")
-    override fun <T> getObject(position: Int):T {
+    override fun <T> getObject(position: Long):T {
         val size = BufferPool.withIntBuffer {
             this.read(it, position)
             it.rewind()
@@ -41,7 +41,7 @@ class EncryptedMemoryMappedStore (filePath: String, context: SchemaContext, dele
      * @return Record id of entity
      * @since 2.2.0
      */
-    override fun writeObject(value: Any?): Int {
+    override fun writeObject(value: Any?): Long {
         val stream = EncryptedBufferStream()
         stream.putObject(value, context)
         stream.flip()
@@ -68,7 +68,7 @@ class EncryptedMemoryMappedStore (filePath: String, context: SchemaContext, dele
      * @since 2.2.0
      * @return Streamable buffer
      */
-    override fun readObject(position: Int, size: Int): BufferStream {
+    override fun readObject(position: Long, size: Int): BufferStream {
         val buffer = read(position, size)
         return EncryptedBufferStream(buffer!!.byteBuffer)
     }
