+++ onyx-database/src/main/kotlin/com/onyx/extension/IManagedEntity$Store.kt
@@ -22,7 +22,7 @@ fun IManagedEntity.getDataFile(context:SchemaContext, descriptor: EntityDescript
  * @param partitionId SystemPartitionEntry primary key
  * @since 2.0.0
  */
-fun IManagedEntity.getDataFile(context: SchemaContext, partitionId:Int = partitionId(context), descriptor: EntityDescriptor = descriptor(context)) = context.getPartitionDataFile(descriptor, partitionId)
+fun IManagedEntity.getDataFile(context: SchemaContext, partitionId:Long = partitionId(context), descriptor: EntityDescriptor = descriptor(context)) = context.getPartitionDataFile(descriptor, partitionId)
 
 /**
  * Get record map of the entities that are stored for this matching entity
@@ -31,4 +31,4 @@ fun IManagedEntity.getDataFile(context: SchemaContext, partitionId:Int = partiti
  * @param partitionId SystemPartitionEntry primary key
  * @since 2.0.0
  */
-fun IManagedEntity.records(context: SchemaContext, partitionId: Int = partitionId(context), descriptor: EntityDescriptor = descriptor(context)):DiskMap<Any, IManagedEntity> = getDataFile(context, partitionId, descriptor).getHashMap(descriptor.identifier!!.type, descriptor.entityClass.name)
+fun IManagedEntity.records(context: SchemaContext, partitionId: Long = partitionId(context), descriptor: EntityDescriptor = descriptor(context)):DiskMap<Any, IManagedEntity> = getDataFile(context, partitionId, descriptor).getHashMap(descriptor.identifier!!.type, descriptor.entityClass.name)
